Refactor Display to Display after ALL Tests
Previously, it would be display tests while running them. This
caused odd looking issues with yielders where it would be
test-blank for a few moments-test

We will instead be introducing a test dialog box while running. We
can also store error output here when we're finished (and an option
to show/hide after tests)