Remove Memory Leaks.
This was EVERYWHERE in the project. My bad for not looking at it sooner.

We've added a print stray nodes button for quick debugging and the only
stray nodes that exist are ones by the editor ourselves that we can't do
anything about but doesn't interrupt anything else (or give us the classic
error on exit).