Refactor Output
Changed to a RichTextLabel from a TextEdit Label. We've also
reduced a number of messages such as running test x, running
method x because this is information that results already tells us.

We also removed a Yield Message because our timer should make it
clear that we are yielding, not having the tests hang.