Number of Changes.
We added our first initial proper boostrapping tests. This
however came with some costs. Namely we had to reverse the decision to
allow any script that extended from WATTest to be run. We should
be able to trust people and typing systems with that issue anyway.

We also mucked around a bit with our results display. Scripts
are now collapsed by default. Methods are not, though you can choose
to expand them to see a range of asserts that may or may not
have context written on them.