Rejig UI-Runner-Wrapper Constructor Logic
Runner now holds a constant link to Yielder instead
of taking it in as data from the root script. This was just
being tramped down to the TestWrapper class anyway.

We also moved the add_child method on Yield from the test
wrapper up to the Runner. We want parents to work as factories
for dependencies to avoid having logic (however minimal) in
the children's constructor