Solve Cache Memory Issue in doubler.object()
When we instanced a doubled script after we changed its
source code the second time, we wouldn't be receiving the
updated script. This is because a copy (with outdated source
code) existed in memory already. That cached copy is what Godot
returned to us.

We tried a few things. For example, freeing the objects from memory
but there wasn't a clean way to do this with reference-counted objects
mid-test because they could still be referenced as temporary variables
within the users test methods.

Instead, we added an internal counter in the doubler that counted
how many object instances were created along with how many files existed
in the temp folder at the time. It should be impossible for scripts to have common
names.

So far it seems to work.