Limit Object() to one use
When we made changes to the modified_source_code cache
we would sometimes add duplicate methods. Sometimes this could be
correct and sometimes it could be wrong.

To avoid confusion, each doubler is now solely responsible for ONE doubled
object and no more. We removed two tests that didn't make any sense with this
change implemented (this could suggest those tests were flaky anyway).

We also removed one part of a test that was covered by another (which
checked if methods not overridden by the double were invoked correctly)