Delay clearing Factory Cache to when test ends
So our factory loaded a number of scripts into memory. Sometimes
this would be parents scripts that we weren't interested in and
were just using to get to the inner class we were interested in.

So we stored all of these "uninteresting" scripts into a cache
that got cleared before we returned the double.

However this caused issues because we had just freed some scripts
from memory, so the next time we wanted to do something like X is Class
we had a test behave erratically (it would usually fail on the first run
because we had cleared it from memory and then preceed to pass on the next runs
because it existed in memory somewhere from the earlier runs).

So we've decided to delay clearing stuff from memory until the test ends (everything
is isolated to the test script so we only care about stuff existing in memory
until the end of it). Hence why we cleared it here.