Doubler returns a Method we act on
Instead of telling the doubler to tell the method
what we want to do, we now get a method back from the doubler
and tell the method what to do directly.

We can possibly further affect this by returning the method
every time we call a control method on it.

double.method("add", STATIC).stub(retval, []).stub(retval, [])