Fix Timer
I think our tests either move to fast for
the timer OR block it somehow. We force
the timer to show the time it took on a call to "end".

We might want to replace this with a progress bar or some other animated feature
so it doesn't look like it hangs.