Remove Validator Class & Test
This script was shrinking. Reluctant to remove it first
for testing reasons however we inlined the test method
into runner and method method into test adaptor. We should
be able to push in input into these in their own unit
tests and then check the results against our input. It is bit
of a roundabout trip in comparison but allows us to check a lot of
implicit information inbetween.