Squash Merge Bootstrapping
This is bit of a complicated merge. Essentially a lot
of the more confusing changes were in search of fixing
a linux crash. The crash was caused by using a base
saved resource to save/load complex data (which was the only
way we could manage stubs etc when it came to static scripts).

Instead we're now adding a const collection on the test double
we created that we append the script director to post-creation (we
can't reassign values in an const collection (e.g array[0] = new_value)
but we can use array.append(value).

This is likely to become invalid come 4.0 but it works for now so that's
all that matters.