Update Base Test Scripts.
We reversed the hierachy. The idea is that
we have our base class be the class that
holds the constructor logic and utility methods whereas
the immediate subclasses are designed as state machines.

We've also inlined a number of const scripts into the constructor
but eventually we would want to modularize this. We could
try a factory/builder pattern but it seems overwrought for something
that requires only one type of test.