Move TestLoader Logic Into Game Context
There was no real reason for us to store this outside of
the game context since it itself wasn't a visual component
or needed by a visual component (such as the result tree).

Now when we select a path or script to test we save it in the
ProjectSettings, run the game and then retrieve it. We're using
active run path as the setting name for the moment but it can be
probably be improved.

Since we no longer need to serialize the data, we decided to remove
the tool keyword and change it from resource to reference. This way
we can instantiate different instances of it for tests.

On top of that since it is no longer part of the configuration file, we
are able to add it to the WAT namespace as WAT.TestLoader