Update Runner _process loop.
We pause process while running tests and don't
unpause it until we're finished. This removes
the need for the running variable.

This should also prevent us from popping off new tests that
don't exist. Hopefully this fixes the segfault error (since my
guess is that we tried to access an array value that wasn't there)