Add Print Debugs
We have a segfault that happens between end() emit_signal("ended")
and WAT.Settings.clear()

My best guess is double-freeing a node. So we're adding a check to
see if it is/isn't deleted before hand. If it segfaults and we
see a false or deleted, then we can be sure trying to free something
that was already freed was the problem.