Add Working Tag OptionButton
Plugin.gd sets a WAT/Tags PoolStringArray value in ProjectSettings
if it doesn't already exist. The OptionButton reads these when pressed
and allows user to set one tag for tests to be run by.

We could probably have a batch-tag run option where we run unique sets
of tagged objects (like those that are "Linux" OR "Windows" scripts to
be run together) but I think this may confuse matters and will leave it
up to popular demand to be implemented.