Add Metadata Inspector Plugin
We will have to changeo our option button
to something else because it keeps wiping
the last tagged value if limit is reaching.

Probably a MenuButton that changes text.

Our main aim is to fix that and to add the ability to delete
tags (we'll probably forego the array sizer in favor of a global
add_tag popup button that adds a label with a thrash button and
a new row.

It won't work exactly like the inspector but as long as it looks better.