Replace Metadata With Resource
We ran into serialization issues with metadata (it
was never being saved). So I've decided to use a resource
instead.

We can't give a script type hint to a dictionary which caused
other issues with caching scripts. So instead we use two arrays that
share data via index (script at X in scripts owns that tags at X in
tags).

Only thing left now is to run some tagged tests.