Add Assert.That
This is a holdover for class based assertions until they can be implemented.

It takes the obj the method is being called on, the string name of the method,
the context. Then it takes a string placeholder that consumes the object (first)
and then the argument array to be formatted.