Remove Super Check From Writer
This was unnecessary because we already check within the doubled method
anyway so no real need to check within the method creating it.

Ideally though I'd like to move the super call out of the script but I can't
really do that without adding an extra method called "super" and even
then I'm not sure.