Create Copies Of Nested Tests To Load
We can't export nested tests themselves from the editor
and then send them to the testrunner in game context. However
we can create a copy that extends from there and load that instead.

Unfortunately our results display reads it as from the cache rather than
from the initial source, so we need to override this somewhow.