Change TestSelector To Use Metadata
Instead of the awkward enum system we had we can instead change it
so that we attach the metadata to the items themselves and that's where
the tests are stored.

To get the most use out of this, we'll need to make our filemanager very smart by
bundling shared tests between parent and subdirs as seperate keys.