Check If Property Is Exported Variable In Doubled Scenes
We were setting a number of other variables, including the scenes
script, which caused problems so now we explicitly check to see if the
variable is exported. We're currently only doing this on the root node
though, so if we have any subscenes we may need to check against those
as well.