Add Basic Results
We are working on a live wire update that
informs users of things happening when they happen. The
idea here is to keep providing feedback, in some cases the
feedback is so fast that it is meaningless but we're hoping
that gamedev is so async/event-driven that it will pay off
the more tests you do.