Add Signals to CSharp Test
All of the signals work fine as of now.

Our biggest pain point is the test case object with unusual/unsafe
access.

We also want to only show a new tree when it is being created. As
well as that we want to add numbers to it as it goes up.