Fix AlexDarigan/WAT#314 for GDScript
As long as the script is in the defined test directory, the validator will pick it up. Previously, any script that extends from WAT.Test but is broken or has no defined test methods will not be properly included in the test. This can be seen as unexpected, since the test already extends WAT.Test, and is in the defined test folder, so it is as explicit as it can be that this script should be included in the test. Thus, the test should fail if there are any syntax errors or 0 tests defined, making it a more useful feature to detect broken tests.